Revert "Webpack"
