--- app/components/app.jsx
@@ -1,12 +1,0 @@
-import React, { Component } from 'react';
-
-export default class App extends Component {
-
-  render() {
-    return (
-      <div className='main'>
-        woiejfwoiafjaoiejf
-      </div>
-    );
-  }
-}
