--- app/index.js
@@ -1,6 +1,0 @@
-import React, { Component } from 'react';
-import { render } from 'react-dom';
-
-import App from './components/app';
-
-render( <App />, document.getElementById('main') );
