--- index.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <title>handout</title>
-  </head>
-  <body>
-    <div id='main'></div>
-    <script src='bundle.js'></script>
-  </body>
-</html>
