+++ package.json
@@ -4,10 +4,8 @@
   "description": "This is a slackbot built on Node.js with an Express back-end that interacts with a React front-end. It was created during the fourth week of my third module at Turing.",
   "main": "server.js",
   "scripts": {
-    "start": "npm build && nodemon prod/server.js",
-    "dev-start": "nodemon prod/server.js",
-    "build": "webpack --config ./webpack.deployment.config.js",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "nodemon server.js"
   },
   "repository": {
     "type": "git",
@@ -28,16 +26,9 @@
     "express": "^4.15.3",
     "knex": "^0.13.0",
     "nodemon": "^1.11.0",
-    "pg": "^6.2.3",
-    "react": "^15.5.4",
-    "react-dom": "^15.5.4"
+    "pg": "^6.2.3"
   },
   "devDependencies": {
-    "babel-cli": "^6.24.1",
-    "babel-core": "^6.24.1",
-    "babel-loader": "^6.2.10",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-react": "^6.24.1",
     "bookshelf": "^0.10.3",
     "bookshelf-modelbase": "^2.10.4",
     "chai": "^4.0.0",
@@ -47,11 +38,6 @@
     "locus": "^2.0.1",
     "mocha": "^3.4.2",
     "nodemon": "^1.11.0",
-    "pg": "^6.2.3",
-    "react-hot-loader": "^1.3.1",
-    "react-scripts": "^0.9.5",
-    "webpack": "^2.6.1",
-    "webpack-dev-middleware": "^1.10.2",
-    "webpack-hot-middleware": "^2.18.0"
+    "pg": "^6.2.3"
   }
 }
