--- webpack.deployment.config.js
@@ -1,31 +1,0 @@
-const path = require('path');
-const webpack = require('webpack');
-
-module.exports = {
-  devtool: 'source-map',
-
-  entry: [
-    './app/index.js'
-  ],
-  output: {
-    path: path.join(__dirname, 'dist'),
-    filename: 'bundle.js',
-    publicPath: '/prod/'
-  },
-  module: {
-    loaders: [
-      {
-        test: /\.(js|jsx)$/,
-        loader: 'babel-loader',
-        include: path.join(__dirname, 'app'),
-        exclude: /node_modules/,
-        query: {
-          presets: ['es2015', 'react']
-        }
-      }
-    ]
-  },
-  resolve: {
-    extensions: ['', '.js', '.jsx', '.json', '.scss', '.css']
-  }
-};
